refactor: extract MacroTargetEditor component and utilities
This commit implements Phase 1 of the MealPrepCalculator refactoring:

1. Create mealPlannerHelpers.js with reusable utilities:
   - calcTotals, calcTotalsRounded for nutrition calculations
   - categorizeIngredient, getCategoryColors for shopping list
   - Constants: MEALS, roundVal, sanitizeFilename, validatePlanForExport

2. Extract MacroTargetEditor.jsx component:
   - Handles calorie target editing with validation
   - Displays macro percentage distribution
   - Integrates with MacroTargetPopover for percentage adjustment
   - Receives all state and handlers as props

3. Update MealPrepCalculator.jsx:
   - Import utilities instead of defining inline
   - Use MacroTargetEditor component instead of inline JSX
   - Reduce component size by ~150 lines
   - Remove duplicate MacroTargetPopover call

Benefits:
- MacroTargetEditor is now independently testable
- Utilities can be reused in other components
- Main component is simplified and more maintainable
- No behavior changes - 100% backward compatible
